% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_bc_config.R
\name{update_bc_config}
\alias{update_bc_config}
\title{modify basic BioCypher configuration to reroute log and output to specified folders}
\usage{
update_bc_config(
  confpath,
  outputDir = tempfile("biocyphOutput"),
  logDir = tempfile("biocyphLogs"),
  tmpstore = tempfile()
)
}
\arguments{
\item{confpath}{character(1) path to a typical BioCypher configuration which by
default would write log and output to current folder}

\item{outputDir}{character(1) path to output}

\item{logDir}{character(1) path to logs}

\item{tmpstore}{path to revised configuration YAML}
}
\description{
modify basic BioCypher configuration to reroute log and output to specified folders
}
